// src/autogen/interfaces/anthology/2022/03/30/Transformer.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface Transformer {
  language: "jq" | "jsonata" | "jsonnet" | "morphism";
  name?: string;
  outputSchema?: string;
  sourceCode: string;
  sourceCodeChecksum: string;
  supportedInputSchemas?: string[];
  [k: string]: unknown;
}